+++ CarrotMarket/CarrotMarket/Week03/ChatModel.swift
@@ -1,13 +1,11 @@
 //
-//  TestModel.swift
+//  ChatModel.swift
 //  CarrotMarket
 //
 //  Created by seozero on 4/26/25.
 //
 
 import UIKit
-
-// MARK: - Chat Model
 
 struct ChatModel {
     let profileImg: UIImage
@@ -35,30 +33,3 @@ extension ChatModel {
     }
 }
 
-// MARK: - Item Model
-
-struct ItemModel {
-    let itemImg: UIImage
-    let name: String
-    let price: String
-    var heartIsSelected: Bool
-}
-
-extension ItemModel {
-    static func dummy() -> [ItemModel] {
-        return [
-            ItemModel(itemImg: .product1, name: "퉁퉁퉁사후르", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product2, name: "오쏘몰", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product3, name: "애플워치", price: "5000", heartIsSelected: true),
-            ItemModel(itemImg: .product4, name: "가방", price: "5000", heartIsSelected: true),
-            ItemModel(itemImg: .product5, name: "헤드셋", price: "5000", heartIsSelected: true),
-            ItemModel(itemImg: .product6, name: "틴트", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product7, name: "카메라", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product8, name: "로션", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product1, name: "락스", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product4, name: "가방", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product5, name: "헤드셋", price: "5000", heartIsSelected: false),
-            ItemModel(itemImg: .product6, name: "틴트", price: "5000", heartIsSelected: false)
-        ]
-    }
-}
